+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 - (Feature) (Gateway) ArangoDB Auth Token
 - (Feature) (Gateway) Dynamic Configuration
 - (Feature) DebugPackage ArangoRoutes
+- (Feature) (Scheduler) Add Status Conditions
 
 ## [1.2.42](https://github.com/arangodb/kube-arangodb/tree/1.2.42) (2024-07-23)
 - (Maintenance) Go 1.22.4 & Kubernetes 1.29.6 libraries